Move consumption component to dedicated module
