+++ vivarium/environment/components/eco_evo/__init__.py
@@ -1,4 +1,5 @@
+from .consumption.component import ConsumptionComponent
 from .spawn.component import SpawnComponent
 from .spawn.controller import SpawnController
 from .spawn.interface import SpawnInterface
-from .component import ConsumptionComponent, EnergyComponent, ReproductionComponent
+from .component import EnergyComponent, ReproductionComponent
